[ADD] 옷장 메인화면 전용 뷰모델 추가
이에 따라 기존의 ClosetViewModel 내용을 ClothAddViewModel 로 이동
